Migrate Blessed app to TypeScript
